+++ relayer-cli/src/commands.rs
@@ -9,15 +9,15 @@ use std::path::PathBuf;
 
 use abscissa_core::{Command, Configurable, FrameworkError, Help, Options, Runnable};
 
-use crate::commands::channel::ChannelCmds;
 use crate::config::Config;
 
 use self::{
     create::CreateCmds, keys::KeysCmd, light::LightCmd, listen::ListenCmd, query::QueryCmd,
-    start::StartCmd, start_multi::StartMultiCmd, tx::TxCmd, version::VersionCmd,
+    start::StartCmd, start_multi::StartMultiCmd, tx::TxCmd, update::UpdateCmds,
+    version::VersionCmd,
 };
 
-mod channel;
+mod cli_utils;
 mod config;
 mod create;
 mod keys;
@@ -27,6 +27,7 @@ mod query;
 mod start;
 mod start_multi;
 mod tx;
+mod update;
 mod version;
 
 /// Default configuration file path
@@ -55,8 +56,14 @@ pub enum CliCmd {
     Light(LightCmd),
 
     /// The `create` subcommand
-    #[options(help = "Create objects on chains")]
+    #[options(help = "Create objects (client, connection, or channel) on chains")]
     Create(CreateCmds),
+
+    /// The `update` subcommand
+    #[options(
+        help = "Update objects on chains. Currently this sub-commands serves only to update clients"
+    )]
+    Update(UpdateCmds),
 
     /// The `start` subcommand
     #[options(help = "Start the relayer")]
@@ -66,10 +73,6 @@ pub enum CliCmd {
     #[options(help = "Start the relayer in multi-channel mode. \
                       Omit the options to pick up connections from the configuration.")]
     StartMulti(StartMultiCmd),
-
-    /// The `channel` subcommand
-    #[options(help = "Channel functionality for managing channels")]
-    Channel(ChannelCmds),
 
     /// The `query` subcommand
     #[options(help = "Query objects from the chain")]